Added GraphEntity as parent of GraphUser and GraphGroup
