+++ azd/src/main/java/org/azd/graph/types/GraphUser.java
@@ -7,19 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * Graph user entity
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GraphUser {
-    /***
-     * This field contains zero or more interesting links about the graph subject.
-     * These links may be invoked to obtain additional relationships or more detailed information about this graph subject.
-     */
-    @JsonProperty("_links")
-    private GraphReferenceLinks _links;
-    /***
-     * The descriptor is the primary way to reference the graph subject while the system is running.
-     * This field will uniquely identify the same graph subject across both Accounts and Organizations.
-     */
-    @JsonProperty("descriptor")
-    private String descriptor;
+public class GraphUser extends GraphEntity {
     /***
      * The short, generally unique name for the user in the backing directory.
      * For AAD users, this corresponds to the mail nickname, which is often but not
@@ -28,16 +16,6 @@ public class GraphUser {
      */
     @JsonProperty("directoryAlias")
     private String directoryAlias;
-    /***
-     * This is the non-unique display name of the graph subject. To change this field, you must alter its value in the source provider.
-     */
-    @JsonProperty("displayName")
-    private String displayName;
-    /***
-     * This represents the name of the container of origin for a graph member.
-     */
-    @JsonProperty("domain")
-    private String domain;
     /***
      * When true, the group has been deleted in the identity provider
      */
@@ -49,61 +27,10 @@ public class GraphUser {
     @JsonProperty("legacyDescriptor")
     private String legacyDescriptor;
     /***
-     * The email address of record for a given graph member. This may be different than the principal name.
-     */
-    @JsonProperty("mailAddress")
-    private String mailAddress;
-    /***
      * The meta type of the user in the origin, such as "member", "guest", etc. See UserMetaType for the set of possible values.
      */
     @JsonProperty("metaType")
     private String metaType;
-    /***
-     * The type of source provider for the origin identifier (ex:AD, AAD, MSA)
-     */
-    @JsonProperty("origin")
-    private String origin;
-    /***
-     * The unique identifier from the system of origin.
-     * Typically a sid, object id or Guid. Linking and unlinking operations
-     * can cause this value to change for a user because the user is not backed by a different provider
-     * and has a different unique id in the new provider.
-     */
-    @JsonProperty("originId")
-    private String originId;
-    /***
-     * This is the PrincipalName of this graph member from the source provider.
-     * The source provider may change this field over time and it is not guaranteed
-     * to be immutable for the life of the graph member by VSTS.
-     */
-    @JsonProperty("principalName")
-    private String principalName;
-    /***
-     * This field identifies the type of the graph subject (ex: Group, Scope, User).
-     */
-    @JsonProperty("subjectKind")
-    private String subjectKind;
-    /***
-     * This url is the full route to the source resource of this graph subject.
-     */
-    @JsonProperty("url")
-    private String url;
-
-    public GraphReferenceLinks get_links() {
-        return _links;
-    }
-
-    public void set_links(GraphReferenceLinks _links) {
-        this._links = _links;
-    }
-
-    public String getDescriptor() {
-        return descriptor;
-    }
-
-    public void setDescriptor(String descriptor) {
-        this.descriptor = descriptor;
-    }
 
     public String getDirectoryAlias() {
         return directoryAlias;
@@ -111,22 +38,6 @@ public class GraphUser {
 
     public void setDirectoryAlias(String directoryAlias) {
         this.directoryAlias = directoryAlias;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public void setDisplayName(String displayName) {
-        this.displayName = displayName;
-    }
-
-    public String getDomain() {
-        return domain;
-    }
-
-    public void setDomain(String domain) {
-        this.domain = domain;
     }
 
     public boolean isDeletedInOrigin() {
@@ -145,14 +56,6 @@ public class GraphUser {
         this.legacyDescriptor = legacyDescriptor;
     }
 
-    public String getMailAddress() {
-        return mailAddress;
-    }
-
-    public void setMailAddress(String mailAddress) {
-        this.mailAddress = mailAddress;
-    }
-
     public String getMetaType() {
         return metaType;
     }
@@ -161,58 +64,18 @@ public class GraphUser {
         this.metaType = metaType;
     }
 
-    public String getOrigin() {
-        return origin;
-    }
-
-    public void setOrigin(String origin) {
-        this.origin = origin;
-    }
-
-    public String getOriginId() {
-        return originId;
-    }
-
-    public void setOriginId(String originId) {
-        this.originId = originId;
-    }
-
-    public String getPrincipalName() {
-        return principalName;
-    }
-
-    public void setPrincipalName(String principalName) {
-        this.principalName = principalName;
-    }
-
-    public String getSubjectKind() {
-        return subjectKind;
-    }
-
-    public void setSubjectKind(String subjectKind) {
-        this.subjectKind = subjectKind;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
     @Override
     public String toString() {
         return "GraphUser{" +
-                "_links=" + _links +
-                ", descriptor='" + descriptor + '\'' +
-                ", directoryAlias='" + directoryAlias + '\'' +
-                ", displayName='" + displayName + '\'' +
-                ", domain='" + domain + '\'' +
+                "directoryAlias='" + directoryAlias + '\'' +
                 ", isDeletedInOrigin=" + isDeletedInOrigin +
                 ", legacyDescriptor='" + legacyDescriptor + '\'' +
-                ", mailAddress='" + mailAddress + '\'' +
                 ", metaType='" + metaType + '\'' +
+                ", _links=" + _links +
+                ", descriptor='" + descriptor + '\'' +
+                ", displayName='" + displayName + '\'' +
+                ", domain='" + domain + '\'' +
+                ", mailAddress='" + mailAddress + '\'' +
                 ", origin='" + origin + '\'' +
                 ", originId='" + originId + '\'' +
                 ", principalName='" + principalName + '\'' +
