+++ azd/src/main/java/org/azd/graph/types/GroupReferenceLinks.java
@@ -7,61 +7,10 @@ import org.azd.common.types.Reference;
 /***
  * This field contains zero or more interesting links about the graph subject.
  * These links may be invoked to obtain additional relationships or more detailed information about this graph subject.
+ * @deprecated This is identical to {@link GraphReferenceLinks} and all references should be updated accordingly
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GroupReferenceLinks {
-    /***
-     * Self reference url
-     */
-    @JsonProperty("self")
-    private Reference self;
-    /***
-     * Represents the ulr of membership API
-     */
-    @JsonProperty("memberships")
-    private Reference memberships;
-    /***
-     * Membership state url
-     */
-    @JsonProperty("membershipState")
-    private Reference membershipState;
-    /***
-     * Storage key url
-     */
-    @JsonProperty("storageKey")
-    private Reference storageKey;
-
-    public Reference getSelf() {
-        return self;
-    }
-
-    public void setSelf(Reference self) {
-        this.self = self;
-    }
-
-    public Reference getMemberships() {
-        return memberships;
-    }
-
-    public void setMemberships(Reference memberships) {
-        this.memberships = memberships;
-    }
-
-    public Reference getMembershipState() {
-        return membershipState;
-    }
-
-    public void setMembershipState(Reference membershipState) {
-        this.membershipState = membershipState;
-    }
-
-    public Reference getStorageKey() {
-        return storageKey;
-    }
-
-    public void setStorageKey(Reference storageKey) {
-        this.storageKey = storageKey;
-    }
+public class GroupReferenceLinks extends GraphReferenceLinks {
 
     @Override
     public String toString() {
